Add read/write tests using the SIO backend
